+++ cppFallingSands-master/src/Elements/BaseElements/Moveable.hpp
@@ -7,18 +7,18 @@
 
 class Moveable : public Pixel {
 public:
-    Moveable();
-    ~Moveable();
-    double randomNumber();
-    void updateVelocity(int &newCol, int upDown);
-    void resetVelocity(){
-        yVelocity = 0;
-    }
-    bool isSpaceFree(Chunk &vec, int row, int col);
+	Moveable();
+	~Moveable();
+	float randomNumber();
+	void updateVelocity(int& newCol, int upDown);
+	void resetVelocity() {
+		yVelocity = 0;
+	}
+	bool isSpaceFree(Chunk& vec, int row, int col);
 private:
-    int yVelocity{0};
-    int xVelocity{0};
-    float verticalToHorizontalRation{0.4};
+	int yVelocity{ 0 };
+	int xVelocity{ 0 };
+	float verticalToHorizontalRation{ 0.4f };
 };
 
 #endif /* MOVEABLE_HPP */
